+++ server/rest/user.rest
@@ -26,8 +26,8 @@ POST http://localhost:3002/api/users/login
 Content-Type: application/json
 
 {
-    "email": "dicken@test.se",
-    "password": "1234"
+  "email": "dicken@test.se",
+  "password": "1234"
 }
 
 ###
