+++ server/src/controllers/user.controller.js
@@ -4,62 +4,74 @@ const bcrypt = require("bcrypt");
 const cors = require("cors");
 const cookieSession = require("cookie-session");
 const app = express();
-const { logIn } = require("./routers/user.routero");
 const User = require("../models/user.model");
-const router = require("./routers/user.router");
 
-
-
-/* app.use(
-  cors({
-    origin: ["http://localhost:3000", "http://localhost:3002"],
-    credentials: true,
-  })
-);
-
-app.use(cookieSession({
-  name: 'GaryLovesCookies',
-  maxAge: 1000 * 1000, // short duration to check easily while developing
-  secret: 'apskda9s8d7236uvjbkajdnfhoias89d70f62t3yhdjhvfuastadcych',
-  httpOnly: true,
-}))
- */
-// login
-const logIn = (logIn) => {
-console.log("log in funtionen");
-  if (req.body.email && req.body.password) {
-    User.authenticate(req.body.email, req.body.password, (err, user) => {
-      if (err) {
-        res.status(401).json({
-          status: "Wrong name"
-        });
-      } else if (user) {
-        //store authentication session
-        req.session.userId = user._id;
-
-        res.status(200).json({
-          status: "Authenticated",
-          user: {
-            _id: user._id,
-            email: user.email,
-          },
-        });
-      } else {
-        res.status(401).json({
-          status: "Wrong password"
-        });
-      }
-    });
+// GET ALL
+const getAllUsers = async (req, res) => {
+  try {
+    const users = await User.find();
+    res.status(200).json(users);
+  } catch (err) {
+    res.status(500).json(err);
   }
-}
+};
+// GET ONE
+const getOneById = async (req, res) => {
+  try {
+    const user = await User.findById(req.params.userId);
+    res.status(200).json(user);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+};
+// CREATE NEW USER
+const newUser = async (req, res) => {
+  const userData = {
+    firstname: req.body.firstname,
+    lastname: req.body.lastname,
+    phonenumber: "",
+    email: req.body.email,
+    adress: "",
+    postcode: "",
+    city: "",
+    creditcardnumber: "",
+    password: req.body.password,
+    role: "customer",
+  };
 
-router.post("/api/users/login", async (req, res) => {
-  await UserModel.find({ name: req.body.name })
+  User.findOne({ email: userData.email }, (err, queriedUser) => {
+    if (err) {
+      console.log("Error finding user in database", err);
+      return;
+    }
+
+    if (!queriedUser) {
+      User.create(userData, (err, user) => {
+        if (err) {
+          console.log(err);
+
+          res.status(400).json(err);
+        } else {
+          // store authentication session
+          // req.session.userId = user._id;
+
+          res.status(201).json({ status: "Authenticated", user });
+        }
+      });
+    } else {
+      res.status(401).json({ status: "User-name already taken" });
+    }
+  });
+};
+
+// LOG IN
+const logInUser = async (req, res) => {
+  await User.find({ email: req.body.email })
     .exec()
     .then((user) => {
       if (user.length < 1) {
         return res.status(401).json({
-          message: "Auth failed",
+          message: "Auth failed at start",
         });
       }
       bcrypt.compare(req.body.password, user[0].password, (err, result) => {
@@ -69,12 +81,12 @@ router.post("/api/users/login", async (req, res) => {
           });
         }
         if (result) {
-          req.session.userId = user[0]._id
+          req.session.userId = user[0]._id;
 
           return res.status(200).json({
             message: "Auth successful",
             userId: user[0]._id,
-            userName: user[0].name,
+            email: user[0].email,
           });
         }
         res.status(401).json({
@@ -87,6 +99,66 @@ router.post("/api/users/login", async (req, res) => {
         error: error,
       });
     });
-});
+};
 
-//module.exports = router;
+// LOGOUT
+
+const logOutUser = (req, res, next) => {
+  if (req.session) {
+    // delete session object
+    req.session.destroy(function (err) {
+      if (err) {
+        return next(err);
+      } else {
+        return res.redirect("/");
+      }
+    });
+  }
+};
+
+// DELETE USER
+const deleteUser = async (req, res) => {
+  try {
+    const removedUser = await User.deleteOne({ _id: req.params.userId });
+    res.status(200).json("User removed");
+  } catch (err) {
+    res.status(500).json(err);
+  }
+};
+
+// UPDATE USER
+const updateUser = async (req, res) => {
+  try {
+    let user = await User.findById(req.params.userId);
+
+    if (user) {
+      user.firstname = user.firstname;
+      user.lastname = user.lastname;
+      user.phonenumber = req.body.phonenumber;
+      user.email = user.email;
+      user.adress = req.body.adress;
+      user.postcode = req.body.postcode;
+      user.city = req.body.city;
+      user.creditcardnumber = req.body.creditcardnumber;
+      user.password = user.password;
+      user.role = "customer";
+      await user.save();
+
+      res.status(200).json(user);
+    } else {
+      res.status(404).json({ status: "User not found in db" });
+    }
+  } catch (err) {
+    res.status(500).json(err);
+  }
+};
+
+module.exports = {
+  getAllUsers,
+  getOneById,
+  newUser,
+  logInUser,
+  logOutUser,
+  deleteUser,
+  updateUser,
+};
